+++ tasks/seq/opolin_d_sum_by_columns/src/ops_seq.cpp
@@ -11,7 +11,7 @@ using namespace std::chrono_literals;
 
 bool opolin_d_sum_by_columns_seq::SumColumnsMatrixSequential::PreProcessingImpl() {
   // init data
-  auto *ptr = reinterpret_cast<double *>(task_data->inputs[0]);
+  auto *ptr = reinterpret_cast<int *>(task_data->inputs[0]);
   input_matrix_.assign(ptr, ptr + (rows_ * cols_));
   output_.resize(cols_, 0.0);
   return true;
@@ -45,7 +45,7 @@ bool opolin_d_sum_by_columns_seq::SumColumnsMatrixSequential::RunImpl() {
 }
 
 bool opolin_d_sum_by_columns_seq::SumColumnsMatrixSequential::PostProcessingImpl() {
-  auto *out = reinterpret_cast<double *>(task_data->outputs[0]);
+  auto *out = reinterpret_cast<int *>(task_data->outputs[0]);
   std::ranges::copy(output_, out);
   return true;
 }
